LSC

Documentation

stage3 :: Compiler' NetGraph

stage4 :: Compiler' NetGraph

globalPlacement :: Compiler' NetGraph

detailedPlacement :: Compiler' NetGraph

fastDP :: Compiler' NetGraph

estimate :: Compiler' NetGraph

finalEstimate :: Compiler' NetGraph

netGraph :: DAG Identifier NetGraph

circuit :: DAG Identifier RTL

newtype Endomorph c a

Constructors

Endomorph

Fields

appEndomorph :: c a a

type Compiler' a = Compiler a a

type Compiler = LSR LS

compiler :: Compiler a b -> a -> LSC b

env :: Confinement () -> Compiler a b -> Compiler a b

remote_ :: LSC b -> Compiler () b

remote :: (a -> LSC b) -> Compiler a b

local_ :: LSC b -> Compiler () b

local :: (a -> LSC b) -> Compiler a b

expensive :: NFData b => (a -> b) -> Compiler a b

type Strategy' a = Strategy a a

type Strategy a b = Compiler a b -> Compiler a b

strategy0 :: Int -> Strategy' a

strategy1 :: Comparison a -> Strategy' a

strategy2 :: Comparison a -> Strategy' a

iterator :: ArrowChoice a => Word -> a b b -> a [b] [b]

iterator1 :: Arrow a => Word -> a b b -> a [b] [b]

newtype LS a b

A Kleisli category with specialized arrow instances for concurrency and exception handling

Constructors

LS

Fields

unLS :: a -> LSC b

data NoResult

Constructors

NoResult

noResult :: LS a b

data LSZero

Constructors

LSZero

newtype LSR ls a b

An arrow transformer with effects on the evaluation of the underlying arrow

Constructors

LSR

Fields

unLSR :: Algebraic ls a b

reduce :: Arrow ls => LSR ls a b -> ls a b