LSC.SegmentTree

Documentation

data SegmentTree a

An implementation closer to the literature

fromList :: Ord a => [(a, a)] -> SegmentTree a

constructSegmentTree :: Ord a => [a] -> [(a, a)] -> SegmentTree a

This is unsafe: constructSegmentTree assumes its first argument to be distinct and in ascending order

endpoints :: Ord a => [(a, a)] -> [a]

maxDensity :: SegmentTree a -> Int

densityRatio :: Ord a => (a, a) -> SegmentTree a -> Rational

densityOn :: Ord a => a -> SegmentTree a -> Int

densityOver :: Ord a => (a, a) -> SegmentTree a -> Int

unsafeDensityRatio :: Ord a => (a, a) -> SegmentTree a -> Rational

unsafeDensityOver :: Ord a => (a, a) -> SegmentTree a -> Int

push :: Ord a => (a, a) -> SegmentTree a -> SegmentTree a

Insert an edge, mnemonic: to push a needle into the haystack

pull :: Ord a => (a, a) -> SegmentTree a -> SegmentTree a

Delete an edge, mnemonic: to pull a needle from the haystack

unsafePush :: Ord a => (a, a) -> SegmentTree a -> SegmentTree a

Insert an edge, mnemonic: to push a needle into the haystack

unsafePull :: Ord a => (a, a) -> SegmentTree a -> SegmentTree a

Delete an edge, mnemonic: to pull a needle from the haystack

showTree :: Show a => SegmentTree a -> String

data Stabs a m

Custom tree data type for creating stabbing queries over monoids, e. g. always appending `Sum 1` counts the number of intervals intersecting the stabbing abscissa.

Note that the underlying monoid type may be changed by using fmap.

skeleton :: (Ord a, Monoid m) => [a] -> Stabs a m

append :: (Ord a, Semigroup m) => (a, a) -> m -> Stabs a m -> Stabs a m

unsafeAppend :: (Ord a, Semigroup m) => (a, a) -> m -> Stabs a m -> Stabs a m

adjust :: Ord a => (a, a) -> (m -> m) -> Stabs a m -> Stabs a m

unsafeAdjust :: Ord a => (a, a) -> (m -> m) -> Stabs a m -> Stabs a m

stab :: (Ord a, Monoid m) => a -> Stabs a m -> m

stabOver :: (Ord a, Monoid m) => (a, a) -> Stabs a m -> m

unsafeStabOver :: (Ord a, Monoid m) => (a, a) -> Stabs a m -> m